pyproject.toml
[tool.poetry]
name = "omspy"
version = "0.0.0"
description = ""
license = "MIT"
authors = ["Ubermensch <uberdeveloper001@gmail.com>"]
readme = "README.md"
repository = "https://github.com/uberdeveloper/omspy"

[tool.poetry.dependencies]
python = "^3.7"
pendulum = "^2.1.2"
pydantic = "^1.8.2"
PyYAML = "^5.4.1"

[tool.poetry.group.test.dependencies]
coverage = "^6.0.2"
selenium = "^3.141.0"
pytest = "^5.2"
pytest-watch = "^4.2.0"

[tool.poetry.group.brokers.dependencies]
pyotp = "^2.6.0"
kiteconnect = "^3.9.4"
requests = "^2.25.0"
fyers-apiv2 = "^2.0.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
